+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### master
+
+- internals: selecting into json or json agg does not generate subselects any more
+
 ### 3.2.0
 
 - fix empty `select`